+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Usage metrics numbers are better formatted (decimal and thousands separators).
+- Don't set fallback size for `ModelBuilder` any longer.
 
 ### Fixed
 
@@ -16,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix CSS issue in `Tabs`: Boxed Tabs were somtimes a bit too small or large.
 - Fix CSS issue in `Logs`: Expand/Collapse arrow changes again based on state.
 - `Process`: Fallback to `unnamed` if no process id is present (to avoid issues with the `ModelBuilder` for example).
+- `ModelBuilder`: Fix issue when parameter name and node id conflict.
 
 ## [2.3.1] - 2020-07-09